requirements.txt
requests
anthropic
python-dotenv
python-telegram-bot
pypdf
python-docx
